refactor: reorganize cache implementation and update imports across modules
