+++ src/install.ts
@@ -8,7 +8,7 @@ import fs from "node:fs";
 import path from "node:path";
 import os from "node:os";
 import got, { RequestError } from "got";
-import { cache } from "./astra.js";
+import { cache } from "./helpers/cache.js";
 export default async function install({ ver }: { ver: string }) {
 	let versionName: string;
 	try {
@@ -32,10 +32,10 @@ export default async function install({ ver }: { ver: string }) {
 				"User-Agent": "AstraCLI",
 			},
 			cache: {
-					get: (key: string) => cache.get(key),
-					set: (key: string, value: unknown) => cache.set(key, value),
-					delete: (key: string) => cache.delete(key),
-					clear: () => cache.clear(),
+				get: (key: string) => cache.get(key),
+				set: (key: string, value: unknown) => cache.set(key, value),
+				delete: (key: string) => cache.delete(key),
+				clear: () => cache.clear(),
 			},
 		},
 	).json()) as Record<string, unknown>;
